% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd_customFunction.R
\name{createCustomRmd}
\alias{createCustomRmd}
\title{Function to create custom rmd file for reprot generation}
\usage{
createCustomRmd(code = "", sectionHash = "#", comment_after = "",
  comment_prior = "", chunkopts = "echo=FALSE,warnings=FALSE,message=FALSE",
  sectionHeader = "R markdown section", chunkVar = "",
  outRmd = "test.Rmd", outRmdDir = ".", append = F, renderOutput = F,
  browse = interactive(), logo = "", projectCode = "QXXXX")
}
\arguments{
\item{code}{R code(in quotes) to produce the desired out. This option takes '\'r (blackslashfollowed by r, without quotes) as line break.}

\item{sectionHash}{hash to indicate subsection # is section 1 and ## subsection of previous section [default #]}

\item{comment_after}{Text to appear after the plot}

\item{comment_prior}{Text to appear above the plot}

\item{sectionHeader}{section title}

\item{chunkVar}{chunck variable name in single quote. This value is used as suffix to generate plot output files.}

\item{outRmd}{Name of the output rmd file}

\item{outRmdDir}{Path to output folder to save all outputs and Rmd file [Default = current working folder]}

\item{append}{A logical value to append existing file or generate new file. [Default =F, generate new file]}

\item{renderOutput}{A logical value to render the generated Rmd to HTML document.}

\item{browse}{A logical value to open the generated HTML document in browser.}

\item{logo}{Path to logo file}

\item{projectCode}{Qbic project code}

\item{chuckopts}{R code chunck options in quotes [Default= "echo=FALSE,warnings=FALSE,message=FALSE"]}
}
\value{
Generate a Rmd file and render it to HTML.
}
\description{
Function to create Rmd file and HTML output
}
\details{
This function isinspried from function in R packgae regionReprot
}
\examples{
 createCustomRmd(code = "plot(1:10,1:10)\\rplot(1:20,1:20)",append = F,outRmd = "Testing.Rmd",outRmdDir  = "../",chunkopts ="echo=FALSE")
 createCustomRmd(code = "for(i in 1:12){\\rprint(i)\\r}",append = T,sectionHeader = "heatmap",comment_after = "heatmap",outRmd = "Testing.Rmd",outRmdDir  = "../")
 createCustomRmd(code="plot(1:100,1:100)",sectionHash = "##",comment_prior = "This is a test",append = T,renderOutput = F,outRmd = "Testing.Rmd",sectionHeader = "subheader",outRmdDir  = "../")
 createCustomRmd(code="",append = T,renderOutput = T,outRmd = "Testing.Rmd",outRmdDir  = "../",projectCode="QXXXX")
}
\author{
Praveen baskaran \email{spemannpraveen@gmail.com}
}
