% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPackages.R
\name{pkgTest}
\alias{pkgTest}
\title{pkgTest}
\usage{
pkgTest(package)
}
\arguments{
\item{package}{Package to be checked}
}
\value{
Checks and install package is not found
}
\description{
Test package is installed if not installs it.
see \link{install.packages} for more details
}
\author{
praveen Baskran \email{spemannpraveen@gmail.com}
}
